removed some sim code, made pid values from int to double

+++ Team 3749 2023/src/main/java/frc/robot/utils/Constants.java
@@ -25,9 +25,9 @@ public final class Constants {
         public static final int stop = 0; // speed 0 = motors stop
 
         //PID values
-        public static final int claw_kP = 0;
-        public static final int claw_kI = 0;
-        public static final int claw_kD = 0;
+        public static final double claw_kP = 0;
+        public static final double claw_kI = 0;
+        public static final double claw_kD = 0;
 
         public static final int setpoint = 0;
